// src/types/createUserType.ts
/* eslint-disable prettier/prettier */
export interface CreateUserType {
  name: string;
  uf: string;
  endereco: string;
  cidade: string;
  rg: string;
  cep: string
  datanasc: string;
  phone: string;
  bairro: string;
  email: string;
  cpf: string;
  valorSelecionado: string;
  password: string;
  confirmPassword: string;
}